% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gorkify.R
\name{gorkagram}
\alias{gorkagram}
\title{#' Draw a Gorkagram.}
\usage{
gorkagram(from, to, what, where, col_from = "blue", col_to = "red",
  col_what = "darkorange2", family = "serif", axis_lwd = 7,
  axis_arrow_size = 0.03, axis_labels_h_offset = 0.025,
  axis_labels_v_offset = 0.2, axis_labels_cex = 1.5,
  axis_labels_font = 2, what_label_v_offset = 0.25,
  what_label_cex = 1.25, what_label_font = 3,
  what_arrow_length = 0.03, what_arrow_gap_lh = 0.015,
  what_arrow_gap_rh = 0.03, what_arrow_head_size = 0.015,
  what_ellipse_lwd = 1.5, what_question_mark_gap = 0.05,
  what_question_cex = 1.25)
}
\arguments{
\item{from, to, what}{Labels (strings) for LH polarity of the axis, the RH
polarity, and the variable.}

\item{where}{A number from 0 to 1 indicating the variable's value
along the axis.}

\item{col_from, col_to, col_what}{Label colors (like \code{\link[graphics]{par}}).}

\item{family, axis_labels_font, what_label_font}{Text font graphical parameters (like \code{\link[graphics]{par}}).}

\item{axis_lwd, what_ellipse_lwd}{Line width graphical parameters (like \code{\link[graphics]{par}}).}

\item{axis_arrow_size, axis_labels_h_offset, axis_labels_v_offset}{Graphical specification for axis in x and y units (0 to 1)}

\item{axis_labels_cex, what_label_cex, what_question_cex}{Text size graphical parameters (like \code{\link[graphics]{par}}).}

\item{what_label_v_offset, what_arrow_length, what_arrow_head_size, what_arrow_gap_lh, what_arrow_gap_rh, what_question_mark_gap}{Graphical specifications of variable ellipse in x and y units (0 to 1)}
}
\value{
a gorkagram, as a side-effect.
}
\description{
A \code{gorkagram} is a graph depicting a value of a variable in
a qualitative semantic differential with labelled polarities. The
depiction of the value is not supposed to be exact, but instead,
in conjunction with its label, to be evocative of a degree of
proximity to a polarity. This type of diagram was prosed in Gorka,
2007 (reference below).
}
\section{References}{

   Gorka, S. (2007) Content and end-state-based alteration in the
   practice of political violence since the end of cold war: the
   difference between the terrorism of the cold war and the terrorism
   of al Qaeda: the rise of the "transcendental terrorist". PhD
   Dissertation, Corvinus University, 166.
}

\examples{
# Reproduce "Diagram Nine" in Gorka, 2007 :
gorkagram(from = "Peacekeeping", to = "Thermonuclear war",
          what = "Terrorism", where = 35/100)

}
